added dragon boss
Dragon currently attacks player like a normal enemy when within attack range, and follows player when not. When dragon is below a threshold health value, he enters stage two and begins flying in certain intervals and doing flying attacks. Stage two definitely needs work and perhaps the height that he flies to as our camera is limited in height so it is hard to see the dragon above you (unless we change the camera)